feat(firmware): extend error_flags and add TX/timeout detection
Extend error_flags from uint8_t to uint16_t to support additional
status indicators:

- ERR_FLAG_TX_FAILURES (0x80): Set after 3+ consecutive batch
  transmission failures. Indicates sensor has data but cannot
  deliver it to the hub.

- ERR_FLAG_HIGH_TIMEOUTS (0x100): Set when >25% of reliable/critical
  messages are timing out (requires minimum 10 message sample size).
  Uses NetworkStats to track timeout rates.

Renamed ERR_FLAG_RADIO_ISSUE to ERR_FLAG_TX_FAILURES for clarity.

Implementation adds consecutive_tx_failures_ counter that increments
on batch ACK failure and resets on success.

+++ src/lora/reliable_messenger.cpp
@@ -87,7 +87,7 @@ bool ReliableMessenger::sendSensorData(uint16_t dst_addr, uint8_t sensor_type, c
 
 bool ReliableMessenger::sendHeartbeat(uint16_t dst_addr, uint32_t uptime_seconds,
                                       uint8_t battery_level, uint8_t signal_strength,
-                                      uint8_t active_sensors, uint8_t error_flags)
+                                      uint8_t active_sensors, uint16_t error_flags)
 {
     uint8_t seq_num = getNextSequenceNumber();
