+++ src/lora/reliable_messenger.h
@@ -139,7 +139,7 @@ public:
      * @return true if heartbeat sent successfully
      */
     bool sendHeartbeat(uint16_t dst_addr, uint32_t uptime_seconds, uint8_t battery_level,
-                       uint8_t signal_strength, uint8_t active_sensors, uint8_t error_flags);
+                       uint8_t signal_strength, uint8_t active_sensors, uint16_t error_flags);
 
     /**
      * @brief Send heartbeat response with current datetime (hub to node)
